Work around bug in JSON::RPC::Legacy::Client
If there is a connection problem from LWB::UserAgent,
JSON::RPC::Legacy::Client creates an JSON object from an empty string
and dies with a message about a JSON problem, thus hiding the
connection problem.

These changes test the status line before deciding if it is a
connection problem or rethrowing the original problem.