feat: add function get_data()
This function returns the data from get_information() in a form that is
easier to use with the JSON-RPC-API call 'cmdb.category.save', for
instance to copy data from an object or a template to another object.

The command line option '--get-data' provides access to this data when
using the script idoit-object.